PTEV2/requirements.txt
pre-commit
ipykernel
pandas
geopandas
numpy
matplotlib
rasterio
zarr
hydra-core
omegaconf
torch==2.4.1
torchvision==0.19.1
scikit-learn
wandb
tqdm
pyarrow
h3
antimeridian
rich
